refactor(ide): extract POD types and helpers into separate modules
Extract the graphql-ide crate's monolithic lib.rs into focused modules:

- types.rs: POD types (Position, Range, FilePath, Location, etc.)
- helpers.rs: Shared utility functions for position conversion,
  range adjustment, and diagnostic conversion

This reduces lib.rs from ~5500 lines to ~4400 lines and improves
code organization by separating concerns. The modular structure
makes the codebase more maintainable and easier to navigate.

Changes:
- Create types.rs with all POD type definitions and their tests
- Create helpers.rs with position/offset utilities and conversion functions
- Update lib.rs to re-export from new modules
- Keep AnalysisHost, Analysis, and IdeDatabase in lib.rs as core API
