+++ crates/graphql-ide/src/helpers.rs
@@ -1,0 +1,801 @@
+//! Shared helper functions for IDE features.
+//!
+//! This module contains utility functions used across multiple IDE features
+//! for position/offset conversion, range adjustment, and type formatting.
+
+use crate::symbol::{
+    find_fragment_definition_range, find_fragment_spreads, find_type_definition_range,
+    find_type_references_in_tree,
+};
+use crate::types::{Position, Range};
+
+/// Convert IDE position to byte offset using `LineIndex`
+pub fn position_to_offset(
+    line_index: &graphql_syntax::LineIndex,
+    position: Position,
+) -> Option<usize> {
+    let line_start = line_index.line_start(position.line as usize)?;
+    Some(line_start + position.character as usize)
+}
+
+/// Convert byte offset to IDE Position using `LineIndex`
+#[allow(clippy::cast_possible_truncation)]
+pub fn offset_to_position(line_index: &graphql_syntax::LineIndex, offset: usize) -> Position {
+    let line = line_index.line_col(offset).0;
+    let line_start = line_index.line_start(line).unwrap_or(0);
+    let character = offset - line_start;
+    Position::new(line as u32, character as u32)
+}
+
+/// Convert byte offset range to IDE Range using `LineIndex`
+pub fn offset_range_to_range(
+    line_index: &graphql_syntax::LineIndex,
+    start_offset: usize,
+    end_offset: usize,
+) -> Range {
+    let start = offset_to_position(line_index, start_offset);
+    let end = offset_to_position(line_index, end_offset);
+    Range::new(start, end)
+}
+
+/// Adjust a position for line offset (used for extracted GraphQL from TypeScript/JavaScript)
+///
+/// When GraphQL is extracted from TypeScript/JavaScript files, the line numbers in the
+/// LSP request are relative to the original file, but we need positions relative to the
+/// extracted GraphQL. This function subtracts the `line_offset` to get the correct position.
+#[allow(clippy::cast_possible_truncation)]
+pub const fn adjust_position_for_line_offset(
+    position: Position,
+    line_offset: u32,
+) -> Option<Position> {
+    if line_offset == 0 {
+        return Some(position);
+    }
+
+    if position.line < line_offset {
+        return None;
+    }
+
+    Some(Position::new(
+        position.line - line_offset,
+        position.character,
+    ))
+}
+
+/// Add line offset to a range (used when returning positions from extracted GraphQL)
+///
+/// When returning positions for document symbols in TypeScript/JavaScript files,
+/// we need to add the `line_offset` to convert from GraphQL-relative positions
+/// back to original file positions.
+pub const fn adjust_range_for_line_offset(range: Range, line_offset: u32) -> Range {
+    if line_offset == 0 {
+        return range;
+    }
+
+    Range::new(
+        Position::new(range.start.line + line_offset, range.start.character),
+        Position::new(range.end.line + line_offset, range.end.character),
+    )
+}
+
+/// Convert analysis Position to IDE Position
+pub const fn convert_position(pos: graphql_analysis::Position) -> Position {
+    Position {
+        line: pos.line,
+        character: pos.character,
+    }
+}
+
+/// Convert analysis `DiagnosticRange` to IDE Range
+pub const fn convert_range(range: graphql_analysis::DiagnosticRange) -> Range {
+    Range {
+        start: convert_position(range.start),
+        end: convert_position(range.end),
+    }
+}
+
+/// Convert analysis Severity to IDE `DiagnosticSeverity`
+pub const fn convert_severity(
+    severity: graphql_analysis::Severity,
+) -> crate::types::DiagnosticSeverity {
+    match severity {
+        graphql_analysis::Severity::Error => crate::types::DiagnosticSeverity::Error,
+        graphql_analysis::Severity::Warning => crate::types::DiagnosticSeverity::Warning,
+        graphql_analysis::Severity::Info => crate::types::DiagnosticSeverity::Information,
+    }
+}
+
+/// Convert analysis Diagnostic to IDE Diagnostic
+pub fn convert_diagnostic(diag: &graphql_analysis::Diagnostic) -> crate::types::Diagnostic {
+    crate::types::Diagnostic {
+        range: convert_range(diag.range),
+        severity: convert_severity(diag.severity),
+        message: diag.message.to_string(),
+        code: diag.code.as_ref().map(ToString::to_string),
+        source: diag.source.to_string(),
+    }
+}
+
+/// Result of finding which block contains a position
+pub struct BlockContext<'a> {
+    /// The syntax tree for the block (or main document)
+    pub tree: &'a apollo_parser::SyntaxTree,
+    /// Line offset to add when returning positions (0 for pure GraphQL files)
+    pub line_offset: u32,
+    /// The block source for building `LineIndex` (None for pure GraphQL files)
+    pub block_source: Option<&'a str>,
+}
+
+/// Find which GraphQL block contains the given position
+///
+/// For pure GraphQL files, returns the main tree with `line_offset` from metadata.
+/// For TS/JS files, finds the block containing the position and returns it with
+/// the appropriate line offset.
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_block_for_position(
+    parse: &graphql_syntax::Parse,
+    position: Position,
+    metadata_line_offset: u32,
+) -> Option<(BlockContext<'_>, Position)> {
+    // If no blocks, this is a pure GraphQL file - use main tree
+    if parse.blocks.is_empty() {
+        let adjusted_pos = adjust_position_for_line_offset(position, metadata_line_offset)?;
+        return Some((
+            BlockContext {
+                tree: &parse.tree,
+                line_offset: metadata_line_offset,
+                block_source: None,
+            },
+            adjusted_pos,
+        ));
+    }
+
+    // For TS/JS files, find which block contains the position
+    for block in &parse.blocks {
+        let block_start_line = block.line as u32;
+        let block_start_col = block.column as u32;
+        let block_lines = block.source.chars().filter(|&c| c == '\n').count() as u32;
+        let block_end_line = block_start_line + block_lines;
+
+        if position.line >= block_start_line && position.line <= block_end_line {
+            let adjusted_line = position.line - block_start_line;
+            let adjusted_col = if adjusted_line == 0 {
+                position.character.saturating_sub(block_start_col)
+            } else {
+                position.character
+            };
+            let adjusted_pos = Position::new(adjusted_line, adjusted_col);
+
+            return Some((
+                BlockContext {
+                    tree: &block.tree,
+                    line_offset: block_start_line,
+                    block_source: Some(&block.source),
+                },
+                adjusted_pos,
+            ));
+        }
+    }
+
+    None
+}
+
+/// Find a fragment definition in a parsed file, handling TS/JS blocks correctly
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_fragment_definition_in_parse(
+    parse: &graphql_syntax::Parse,
+    fragment_name: &str,
+    content: graphql_db::FileContent,
+    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
+    metadata_line_offset: u32,
+) -> Option<Range> {
+    if parse.blocks.is_empty() {
+        if let Some((start_offset, end_offset)) =
+            find_fragment_definition_range(&parse.tree, fragment_name)
+        {
+            let file_line_index = graphql_syntax::line_index(db, content);
+            let range = offset_range_to_range(&file_line_index, start_offset, end_offset);
+            return Some(adjust_range_for_line_offset(range, metadata_line_offset));
+        }
+        return None;
+    }
+
+    for block in &parse.blocks {
+        if let Some((start_offset, end_offset)) =
+            find_fragment_definition_range(&block.tree, fragment_name)
+        {
+            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
+            let range = offset_range_to_range(&block_line_index, start_offset, end_offset);
+            return Some(adjust_range_for_line_offset(range, block.line as u32));
+        }
+    }
+
+    None
+}
+
+/// Find a type definition in a parsed file, handling TS/JS blocks correctly
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_type_definition_in_parse(
+    parse: &graphql_syntax::Parse,
+    type_name: &str,
+    content: graphql_db::FileContent,
+    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
+    metadata_line_offset: u32,
+) -> Option<Range> {
+    if parse.blocks.is_empty() {
+        if let Some((start_offset, end_offset)) = find_type_definition_range(&parse.tree, type_name)
+        {
+            let file_line_index = graphql_syntax::line_index(db, content);
+            let range = offset_range_to_range(&file_line_index, start_offset, end_offset);
+            return Some(adjust_range_for_line_offset(range, metadata_line_offset));
+        }
+        return None;
+    }
+
+    for block in &parse.blocks {
+        if let Some((start_offset, end_offset)) = find_type_definition_range(&block.tree, type_name)
+        {
+            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
+            let range = offset_range_to_range(&block_line_index, start_offset, end_offset);
+            return Some(adjust_range_for_line_offset(range, block.line as u32));
+        }
+    }
+
+    None
+}
+
+/// Find all fragment spreads in a parsed file, handling TS/JS blocks correctly
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_fragment_spreads_in_parse(
+    parse: &graphql_syntax::Parse,
+    fragment_name: &str,
+    content: graphql_db::FileContent,
+    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
+    metadata_line_offset: u32,
+) -> Vec<Range> {
+    let mut results = Vec::new();
+
+    if parse.blocks.is_empty() {
+        if let Some(offsets) = find_fragment_spreads(&parse.tree, fragment_name) {
+            let file_line_index = graphql_syntax::line_index(db, content);
+            for offset in offsets {
+                let end_offset = offset + fragment_name.len();
+                let range = offset_range_to_range(&file_line_index, offset, end_offset);
+                results.push(adjust_range_for_line_offset(range, metadata_line_offset));
+            }
+        }
+        return results;
+    }
+
+    for block in &parse.blocks {
+        if let Some(offsets) = find_fragment_spreads(&block.tree, fragment_name) {
+            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
+            for offset in offsets {
+                let end_offset = offset + fragment_name.len();
+                let range = offset_range_to_range(&block_line_index, offset, end_offset);
+                results.push(adjust_range_for_line_offset(range, block.line as u32));
+            }
+        }
+    }
+
+    results
+}
+
+/// Find all type references in a parsed file, handling TS/JS blocks correctly
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_type_references_in_parse(
+    parse: &graphql_syntax::Parse,
+    type_name: &str,
+    content: graphql_db::FileContent,
+    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
+    metadata_line_offset: u32,
+) -> Vec<Range> {
+    let mut results = Vec::new();
+
+    if parse.blocks.is_empty() {
+        if let Some(offsets) = find_type_references_in_tree(&parse.tree, type_name) {
+            let file_line_index = graphql_syntax::line_index(db, content);
+            for offset in offsets {
+                let end_offset = offset + type_name.len();
+                let range = offset_range_to_range(&file_line_index, offset, end_offset);
+                results.push(adjust_range_for_line_offset(range, metadata_line_offset));
+            }
+        }
+        return results;
+    }
+
+    for block in &parse.blocks {
+        if let Some(offsets) = find_type_references_in_tree(&block.tree, type_name) {
+            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
+            for offset in offsets {
+                let end_offset = offset + type_name.len();
+                let range = offset_range_to_range(&block_line_index, offset, end_offset);
+                results.push(adjust_range_for_line_offset(range, block.line as u32));
+            }
+        }
+    }
+
+    results
+}
+
+/// Find field usages in a parsed file that match the given type and field name
+#[allow(clippy::cast_possible_truncation)]
+pub fn find_field_usages_in_parse(
+    parse: &graphql_syntax::Parse,
+    type_name: &str,
+    field_name: &str,
+    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
+    content: graphql_db::FileContent,
+    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
+    metadata_line_offset: u32,
+) -> Vec<Range> {
+    let mut results = Vec::new();
+
+    if parse.blocks.is_empty() {
+        let file_line_index = graphql_syntax::line_index(db, content);
+        let ranges = find_field_usages_in_tree(&parse.tree, type_name, field_name, schema_types);
+        for (start, end) in ranges {
+            let range = offset_range_to_range(&file_line_index, start, end);
+            results.push(adjust_range_for_line_offset(range, metadata_line_offset));
+        }
+        return results;
+    }
+
+    for block in &parse.blocks {
+        let block_line_index = graphql_syntax::LineIndex::new(&block.source);
+        let ranges = find_field_usages_in_tree(&block.tree, type_name, field_name, schema_types);
+        for (start, end) in ranges {
+            let range = offset_range_to_range(&block_line_index, start, end);
+            results.push(adjust_range_for_line_offset(range, block.line as u32));
+        }
+    }
+
+    results
+}
+
+/// Check if `current_type` matches `target_type` directly or implements it as an interface
+fn type_matches_or_implements(
+    current_type: &str,
+    target_type: &str,
+    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
+) -> bool {
+    if current_type == target_type {
+        return true;
+    }
+    if let Some(type_def) = schema_types.get(current_type) {
+        type_def
+            .implements
+            .iter()
+            .any(|i| i.as_ref() == target_type)
+    } else {
+        false
+    }
+}
+
+/// Find all field usages in a tree that match the given type and field name
+#[allow(clippy::too_many_lines)]
+pub fn find_field_usages_in_tree(
+    tree: &apollo_parser::SyntaxTree,
+    target_type: &str,
+    target_field: &str,
+    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
+) -> Vec<(usize, usize)> {
+    use apollo_parser::cst::{CstNode, Definition, Selection};
+
+    fn search_selection_set(
+        selection_set: &apollo_parser::cst::SelectionSet,
+        current_type: &str,
+        target_type: &str,
+        target_field: &str,
+        schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
+        results: &mut Vec<(usize, usize)>,
+    ) {
+        for selection in selection_set.selections() {
+            match selection {
+                Selection::Field(field) => {
+                    if let Some(name) = field.name() {
+                        let field_name = name.text();
+
+                        if type_matches_or_implements(current_type, target_type, schema_types)
+                            && field_name == target_field
+                        {
+                            let range = name.syntax().text_range();
+                            results.push((range.start().into(), range.end().into()));
+                        }
+
+                        if let Some(nested) = field.selection_set() {
+                            if let Some(type_def) = schema_types.get(current_type) {
+                                if let Some(field_def) = type_def
+                                    .fields
+                                    .iter()
+                                    .find(|f| f.name.as_ref() == field_name)
+                                {
+                                    let field_type = field_def.type_ref.name.as_ref();
+                                    search_selection_set(
+                                        &nested,
+                                        field_type,
+                                        target_type,
+                                        target_field,
+                                        schema_types,
+                                        results,
+                                    );
+                                }
+                            }
+                        }
+                    }
+                }
+                Selection::InlineFragment(inline_frag) => {
+                    let fragment_type = inline_frag
+                        .type_condition()
+                        .and_then(|tc| tc.named_type())
+                        .and_then(|nt| nt.name())
+                        .map_or_else(|| current_type.to_string(), |n| n.text().to_string());
+
+                    if let Some(nested) = inline_frag.selection_set() {
+                        search_selection_set(
+                            &nested,
+                            &fragment_type,
+                            target_type,
+                            target_field,
+                            schema_types,
+                            results,
+                        );
+                    }
+                }
+                Selection::FragmentSpread(_) => {}
+            }
+        }
+    }
+
+    let mut results = Vec::new();
+    let doc = tree.document();
+
+    for definition in doc.definitions() {
+        match definition {
+            Definition::OperationDefinition(op) => {
+                let root_type = match op.operation_type() {
+                    Some(op_type) if op_type.mutation_token().is_some() => "Mutation",
+                    Some(op_type) if op_type.subscription_token().is_some() => "Subscription",
+                    _ => "Query",
+                };
+
+                if let Some(selection_set) = op.selection_set() {
+                    search_selection_set(
+                        &selection_set,
+                        root_type,
+                        target_type,
+                        target_field,
+                        schema_types,
+                        &mut results,
+                    );
+                }
+            }
+            Definition::FragmentDefinition(frag) => {
+                let fragment_type = frag
+                    .type_condition()
+                    .and_then(|tc| tc.named_type())
+                    .and_then(|nt| nt.name())
+                    .map(|n| n.text().to_string());
+
+                if let (Some(fragment_type), Some(selection_set)) =
+                    (fragment_type, frag.selection_set())
+                {
+                    search_selection_set(
+                        &selection_set,
+                        &fragment_type,
+                        target_type,
+                        target_field,
+                        schema_types,
+                        &mut results,
+                    );
+                }
+            }
+            _ => {}
+        }
+    }
+
+    results
+}
+
+/// Find variable definition in an operation by name
+pub fn find_variable_definition_in_tree(
+    tree: &apollo_parser::SyntaxTree,
+    var_name: &str,
+    line_index: &graphql_syntax::LineIndex,
+    line_offset: u32,
+) -> Option<Range> {
+    use apollo_parser::cst::{CstNode, Definition};
+
+    let doc = tree.document();
+    for definition in doc.definitions() {
+        if let Definition::OperationDefinition(op) = definition {
+            if let Some(var_defs) = op.variable_definitions() {
+                for var_def in var_defs.variable_definitions() {
+                    if let Some(variable) = var_def.variable() {
+                        if let Some(name) = variable.name() {
+                            if name.text() == var_name {
+                                let range = name.syntax().text_range();
+                                let start: usize = range.start().into();
+                                let end: usize = range.end().into();
+                                let pos_range = offset_range_to_range(line_index, start, end);
+                                return Some(adjust_range_for_line_offset(pos_range, line_offset));
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+    None
+}
+
+/// Find operation definition by name
+pub fn find_operation_definition_in_tree(
+    tree: &apollo_parser::SyntaxTree,
+    op_name: &str,
+    line_index: &graphql_syntax::LineIndex,
+    line_offset: u32,
+) -> Option<Range> {
+    use apollo_parser::cst::{CstNode, Definition};
+
+    let doc = tree.document();
+    for definition in doc.definitions() {
+        if let Definition::OperationDefinition(op) = definition {
+            if let Some(name) = op.name() {
+                if name.text() == op_name {
+                    let range = name.syntax().text_range();
+                    let start: usize = range.start().into();
+                    let end: usize = range.end().into();
+                    let pos_range = offset_range_to_range(line_index, start, end);
+                    return Some(adjust_range_for_line_offset(pos_range, line_offset));
+                }
+            }
+        }
+    }
+    None
+}
+
+/// Find argument definition in schema type's field
+pub fn find_argument_definition_in_tree(
+    tree: &apollo_parser::SyntaxTree,
+    type_name: &str,
+    field_name: &str,
+    arg_name: &str,
+    line_index: &graphql_syntax::LineIndex,
+    line_offset: u32,
+) -> Option<Range> {
+    use apollo_parser::cst::{CstNode, Definition};
+
+    let doc = tree.document();
+    for definition in doc.definitions() {
+        let (name_node, fields_def) = match &definition {
+            Definition::ObjectTypeDefinition(obj) => (obj.name(), obj.fields_definition()),
+            Definition::InterfaceTypeDefinition(iface) => (iface.name(), iface.fields_definition()),
+            _ => continue,
+        };
+
+        let Some(name) = name_node else { continue };
+        if name.text() != type_name {
+            continue;
+        }
+
+        let Some(fields) = fields_def else { continue };
+        for field in fields.field_definitions() {
+            let Some(fname) = field.name() else { continue };
+            if fname.text() != field_name {
+                continue;
+            }
+
+            if let Some(args_def) = field.arguments_definition() {
+                for input_val in args_def.input_value_definitions() {
+                    if let Some(aname) = input_val.name() {
+                        if aname.text() == arg_name {
+                            let range = aname.syntax().text_range();
+                            let start: usize = range.start().into();
+                            let end: usize = range.end().into();
+                            let pos_range = offset_range_to_range(line_index, start, end);
+                            return Some(adjust_range_for_line_offset(pos_range, line_offset));
+                        }
+                    }
+                }
+            }
+        }
+    }
+    None
+}
+
+/// Find the field name at a given offset (for argument context)
+pub fn find_field_name_at_offset(
+    tree: &apollo_parser::SyntaxTree,
+    byte_offset: usize,
+) -> Option<String> {
+    use apollo_parser::cst::{CstNode, Definition, Selection};
+
+    fn check_selection_set(
+        selection_set: &apollo_parser::cst::SelectionSet,
+        byte_offset: usize,
+    ) -> Option<String> {
+        for selection in selection_set.selections() {
+            if let Selection::Field(field) = selection {
+                let range = field.syntax().text_range();
+                let start: usize = range.start().into();
+                let end: usize = range.end().into();
+
+                if byte_offset >= start && byte_offset <= end {
+                    if let Some(args) = field.arguments() {
+                        let args_range = args.syntax().text_range();
+                        let args_start: usize = args_range.start().into();
+                        let args_end: usize = args_range.end().into();
+                        if byte_offset >= args_start && byte_offset <= args_end {
+                            return field.name().map(|n| n.text().to_string());
+                        }
+                    }
+
+                    if let Some(nested) = field.selection_set() {
+                        if let Some(name) = check_selection_set(&nested, byte_offset) {
+                            return Some(name);
+                        }
+                    }
+                }
+            }
+        }
+        None
+    }
+
+    let doc = tree.document();
+    for definition in doc.definitions() {
+        match definition {
+            Definition::OperationDefinition(op) => {
+                if let Some(selection_set) = op.selection_set() {
+                    if let Some(name) = check_selection_set(&selection_set, byte_offset) {
+                        return Some(name);
+                    }
+                }
+            }
+            Definition::FragmentDefinition(frag) => {
+                if let Some(selection_set) = frag.selection_set() {
+                    if let Some(name) = check_selection_set(&selection_set, byte_offset) {
+                        return Some(name);
+                    }
+                }
+            }
+            _ => {}
+        }
+    }
+    None
+}
+
+/// Unwrap a `TypeRef` to get just the base type name (without List or `NonNull` wrappers)
+#[must_use]
+pub fn unwrap_type_to_name(type_ref: &graphql_hir::TypeRef) -> String {
+    type_ref.name.to_string()
+}
+
+/// Format a type reference for display (e.g., "[String!]!")
+pub fn format_type_ref(type_ref: &graphql_hir::TypeRef) -> String {
+    let mut result = type_ref.name.to_string();
+
+    if type_ref.is_list {
+        result = format!("[{result}]");
+        if type_ref.inner_non_null {
+            result = format!("[{}!]", type_ref.name);
+        }
+    }
+
+    if type_ref.is_non_null {
+        result.push('!');
+    }
+
+    result
+}
+
+/// Convert a filesystem path to a file:// URI
+pub fn path_to_file_uri(path: &std::path::Path) -> String {
+    let path_str = path.to_string_lossy();
+
+    if path_str.starts_with("file://") || path_str.contains("://") {
+        return path_str.to_string();
+    }
+
+    if path_str.starts_with('/') {
+        return format!("file://{path_str}");
+    }
+
+    path_str.to_string()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_position_to_offset_helper() {
+        let text = "line 1\nline 2\nline 3";
+        let line_index = graphql_syntax::LineIndex::new(text);
+
+        assert_eq!(
+            position_to_offset(&line_index, Position::new(0, 0)),
+            Some(0)
+        );
+        assert_eq!(
+            position_to_offset(&line_index, Position::new(0, 5)),
+            Some(5)
+        );
+        assert_eq!(
+            position_to_offset(&line_index, Position::new(1, 0)),
+            Some(7)
+        );
+        assert_eq!(
+            position_to_offset(&line_index, Position::new(1, 3)),
+            Some(10)
+        );
+        assert_eq!(
+            position_to_offset(&line_index, Position::new(2, 0)),
+            Some(14)
+        );
+    }
+
+    #[test]
+    fn test_conversion_position() {
+        let analysis_pos = graphql_analysis::Position::new(10, 20);
+        let ide_pos = convert_position(analysis_pos);
+
+        assert_eq!(ide_pos.line, 10);
+        assert_eq!(ide_pos.character, 20);
+    }
+
+    #[test]
+    fn test_conversion_range() {
+        let analysis_range = graphql_analysis::DiagnosticRange::new(
+            graphql_analysis::Position::new(1, 5),
+            graphql_analysis::Position::new(1, 10),
+        );
+        let ide_range = convert_range(analysis_range);
+
+        assert_eq!(ide_range.start.line, 1);
+        assert_eq!(ide_range.start.character, 5);
+        assert_eq!(ide_range.end.line, 1);
+        assert_eq!(ide_range.end.character, 10);
+    }
+
+    #[test]
+    fn test_conversion_severity() {
+        assert_eq!(
+            convert_severity(graphql_analysis::Severity::Error),
+            crate::types::DiagnosticSeverity::Error
+        );
+        assert_eq!(
+            convert_severity(graphql_analysis::Severity::Warning),
+            crate::types::DiagnosticSeverity::Warning
+        );
+        assert_eq!(
+            convert_severity(graphql_analysis::Severity::Info),
+            crate::types::DiagnosticSeverity::Information
+        );
+    }
+
+    #[test]
+    fn test_conversion_diagnostic() {
+        let analysis_diag = graphql_analysis::Diagnostic::with_source_and_code(
+            graphql_analysis::Severity::Warning,
+            "Test warning message",
+            graphql_analysis::DiagnosticRange::new(
+                graphql_analysis::Position::new(2, 0),
+                graphql_analysis::Position::new(2, 10),
+            ),
+            "test-source",
+            "TEST001",
+        );
+
+        let ide_diag = convert_diagnostic(&analysis_diag);
+
+        assert_eq!(ide_diag.severity, crate::types::DiagnosticSeverity::Warning);
+        assert_eq!(ide_diag.message, "Test warning message");
+        assert_eq!(ide_diag.source, "test-source");
+        assert_eq!(ide_diag.code, Some("TEST001".to_string()));
+        assert_eq!(ide_diag.range.start.line, 2);
+        assert_eq!(ide_diag.range.start.character, 0);
+        assert_eq!(ide_diag.range.end.line, 2);
+        assert_eq!(ide_diag.range.end.character, 10);
+    }
+}
