+++ crates/graphql-ide/src/lib.rs
@@ -46,31 +46,38 @@ use salsa::Setter;
 mod file_registry;
 pub use file_registry::FileRegistry;
 
-mod symbol;
+// New modular structure
+mod helpers;
+pub(crate) mod symbol;
+mod types;
+
+// Re-export types from the types module
+pub use types::{
+    CompletionItem, CompletionKind, Diagnostic, DiagnosticSeverity, DocumentSymbol, FilePath,
+    HoverResult, InsertTextFormat, Location, Position, Range, SchemaStats, SymbolKind,
+    WorkspaceSymbol,
+};
+
+// Re-export helpers for internal use
+use helpers::{
+    adjust_range_for_line_offset, convert_diagnostic, find_argument_definition_in_tree,
+    find_block_for_position, find_field_name_at_offset, find_field_usages_in_parse,
+    find_fragment_definition_in_parse, find_fragment_spreads_in_parse,
+    find_operation_definition_in_tree, find_type_definition_in_parse,
+    find_type_references_in_parse, find_variable_definition_in_tree, format_type_ref,
+    offset_range_to_range, path_to_file_uri, position_to_offset,
+};
+// Re-export for use in symbol module
+pub use helpers::unwrap_type_to_name;
+
 use symbol::{
     extract_all_definitions, find_field_definition_full_range, find_fragment_definition_full_range,
-    find_fragment_definition_range, find_fragment_spreads, find_operation_definition_ranges,
-    find_parent_type_at_offset, find_schema_field_parent_type, find_symbol_at_offset,
-    find_type_definition_full_range, find_type_definition_range, find_type_references_in_tree,
-    is_in_selection_set, Symbol,
+    find_operation_definition_ranges, find_parent_type_at_offset, find_schema_field_parent_type,
+    find_symbol_at_offset, find_type_definition_full_range, is_in_selection_set, Symbol,
 };
 
 // Re-export database types that IDE layer needs
 pub use graphql_db::FileKind;
-
-/// Position in a file (editor coordinates, 0-indexed)
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Position {
-    pub line: u32,
-    pub character: u32,
-}
-
-impl Position {
-    #[must_use]
-    pub const fn new(line: u32, character: u32) -> Self {
-        Self { line, character }
-    }
-}
 
 #[cfg(test)]
 /// Helper for tests: extracts cursor position from a string with a `*` marker.
@@ -117,371 +124,7 @@ fn extract_cursor(input: &str) -> (String, Position) {
     (result, Position::new(line, character))
 }
 
-/// Range in a file (editor coordinates)
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Range {
-    pub start: Position,
-    pub end: Position,
-}
-
-impl Range {
-    #[must_use]
-    pub const fn new(start: Position, end: Position) -> Self {
-        Self { start, end }
-    }
-}
-
-/// File path (can be URI or file path)
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct FilePath(pub String);
-
-impl FilePath {
-    pub fn new(path: impl Into<String>) -> Self {
-        Self(path.into())
-    }
-
-    #[must_use]
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<String> for FilePath {
-    fn from(s: String) -> Self {
-        Self(s)
-    }
-}
-
-impl From<&str> for FilePath {
-    fn from(s: &str) -> Self {
-        Self(s.to_string())
-    }
-}
-
-/// Convert a filesystem path to a file:// URI
-///
-/// This handles the common case of absolute Unix paths.
-/// If the path is already a URI (starts with a scheme), it's returned as-is.
-fn path_to_file_uri(path: &std::path::Path) -> String {
-    let path_str = path.to_string_lossy();
-
-    // Already a URI
-    if path_str.starts_with("file://") || path_str.contains("://") {
-        return path_str.to_string();
-    }
-
-    // Absolute Unix path
-    if path_str.starts_with('/') {
-        return format!("file://{path_str}");
-    }
-
-    // For other cases (relative paths, Windows paths), just use as-is
-    // A more complete implementation would handle Windows drive letters
-    path_str.to_string()
-}
-
-/// Location in a specific file
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Location {
-    pub file: FilePath,
-    pub range: Range,
-}
-
-impl Location {
-    #[must_use]
-    pub const fn new(file: FilePath, range: Range) -> Self {
-        Self { file, range }
-    }
-}
-
-/// Completion item kind
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CompletionKind {
-    Field,
-    Type,
-    Fragment,
-    Directive,
-    EnumValue,
-    Argument,
-    Variable,
-}
-
-/// Insert text format for completion items
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum InsertTextFormat {
-    PlainText,
-    Snippet,
-}
-
-/// Completion item
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CompletionItem {
-    pub label: String,
-    pub kind: CompletionKind,
-    pub detail: Option<String>,
-    pub documentation: Option<String>,
-    pub insert_text: Option<String>,
-    pub insert_text_format: Option<InsertTextFormat>,
-    pub sort_text: Option<String>,
-    pub deprecated: bool,
-}
-
-impl CompletionItem {
-    pub fn new(label: impl Into<String>, kind: CompletionKind) -> Self {
-        Self {
-            label: label.into(),
-            kind,
-            detail: None,
-            documentation: None,
-            insert_text: None,
-            insert_text_format: None,
-            sort_text: None,
-            deprecated: false,
-        }
-    }
-
-    #[must_use]
-    pub fn with_detail(mut self, detail: impl Into<String>) -> Self {
-        self.detail = Some(detail.into());
-        self
-    }
-
-    #[must_use]
-    pub fn with_documentation(mut self, doc: impl Into<String>) -> Self {
-        self.documentation = Some(doc.into());
-        self
-    }
-
-    #[must_use]
-    pub fn with_insert_text(mut self, text: impl Into<String>) -> Self {
-        self.insert_text = Some(text.into());
-        self
-    }
-
-    #[must_use]
-    pub const fn with_insert_text_format(mut self, format: InsertTextFormat) -> Self {
-        self.insert_text_format = Some(format);
-        self
-    }
-
-    #[must_use]
-    pub fn with_sort_text(mut self, sort_text: impl Into<String>) -> Self {
-        self.sort_text = Some(sort_text.into());
-        self
-    }
-
-    #[must_use]
-    pub const fn with_deprecated(mut self, deprecated: bool) -> Self {
-        self.deprecated = deprecated;
-        self
-    }
-}
-
-/// Hover information
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct HoverResult {
-    /// Markdown content
-    pub contents: String,
-    /// Optional range for the hover
-    pub range: Option<Range>,
-}
-
-impl HoverResult {
-    pub fn new(contents: impl Into<String>) -> Self {
-        Self {
-            contents: contents.into(),
-            range: None,
-        }
-    }
-
-    #[must_use]
-    pub const fn with_range(mut self, range: Range) -> Self {
-        self.range = Some(range);
-        self
-    }
-}
-
-/// Diagnostic severity
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DiagnosticSeverity {
-    Error,
-    Warning,
-    Information,
-    Hint,
-}
-
-/// Diagnostic (error, warning, hint)
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Diagnostic {
-    pub range: Range,
-    pub severity: DiagnosticSeverity,
-    pub message: String,
-    pub code: Option<String>,
-    pub source: String,
-}
-
-impl Diagnostic {
-    pub fn new(
-        range: Range,
-        severity: DiagnosticSeverity,
-        message: impl Into<String>,
-        source: impl Into<String>,
-    ) -> Self {
-        Self {
-            range,
-            severity,
-            message: message.into(),
-            code: None,
-            source: source.into(),
-        }
-    }
-
-    #[must_use]
-    pub fn with_code(mut self, code: impl Into<String>) -> Self {
-        self.code = Some(code.into());
-        self
-    }
-}
-
-/// Kind of GraphQL symbol for document/workspace symbols
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SymbolKind {
-    /// Object type definition
-    Type,
-    /// Field definition
-    Field,
-    /// Query operation
-    Query,
-    /// Mutation operation
-    Mutation,
-    /// Subscription operation
-    Subscription,
-    /// Fragment definition
-    Fragment,
-    /// Enum value
-    EnumValue,
-    /// Scalar type
-    Scalar,
-    /// Input type
-    Input,
-    /// Interface type
-    Interface,
-    /// Union type
-    Union,
-    /// Enum type
-    Enum,
-}
-
-/// A document symbol (hierarchical structure for outline view)
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DocumentSymbol {
-    /// Symbol name
-    pub name: String,
-    /// Symbol kind
-    pub kind: SymbolKind,
-    /// Optional detail (e.g., type signature)
-    pub detail: Option<String>,
-    /// Full range of the symbol (entire definition)
-    pub range: Range,
-    /// Selection range (just the name)
-    pub selection_range: Range,
-    /// Child symbols (e.g., fields within a type)
-    pub children: Vec<DocumentSymbol>,
-}
-
-impl DocumentSymbol {
-    pub fn new(
-        name: impl Into<String>,
-        kind: SymbolKind,
-        range: Range,
-        selection_range: Range,
-    ) -> Self {
-        Self {
-            name: name.into(),
-            kind,
-            detail: None,
-            range,
-            selection_range,
-            children: Vec::new(),
-        }
-    }
-
-    #[must_use]
-    pub fn with_detail(mut self, detail: impl Into<String>) -> Self {
-        self.detail = Some(detail.into());
-        self
-    }
-
-    #[must_use]
-    pub fn with_children(mut self, children: Vec<Self>) -> Self {
-        self.children = children;
-        self
-    }
-}
-
-/// A workspace symbol (flat structure for global search)
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WorkspaceSymbol {
-    /// Symbol name
-    pub name: String,
-    /// Symbol kind
-    pub kind: SymbolKind,
-    /// Location of the symbol
-    pub location: Location,
-    /// Container name (e.g., parent type for fields)
-    pub container_name: Option<String>,
-}
-
-impl WorkspaceSymbol {
-    pub fn new(name: impl Into<String>, kind: SymbolKind, location: Location) -> Self {
-        Self {
-            name: name.into(),
-            kind,
-            location,
-            container_name: None,
-        }
-    }
-
-    #[must_use]
-    pub fn with_container(mut self, container: impl Into<String>) -> Self {
-        self.container_name = Some(container.into());
-        self
-    }
-}
-
-/// Statistics about schema types
-#[derive(Debug, Clone, Default)]
-pub struct SchemaStats {
-    /// Number of object types
-    pub objects: usize,
-    /// Number of interface types
-    pub interfaces: usize,
-    /// Number of union types
-    pub unions: usize,
-    /// Number of enum types
-    pub enums: usize,
-    /// Number of scalar types
-    pub scalars: usize,
-    /// Number of input object types
-    pub input_objects: usize,
-    /// Total number of fields across all types
-    pub total_fields: usize,
-    /// Number of directive definitions
-    pub directives: usize,
-}
-
-impl SchemaStats {
-    /// Total number of types (all kinds)
-    #[must_use]
-    pub fn total_types(&self) -> usize {
-        self.objects
-            + self.interfaces
-            + self.unions
-            + self.enums
-            + self.scalars
-            + self.input_objects
-    }
-}
+// POD types are now defined in types.rs and re-exported above
 
 /// Input: Lint configuration
 ///
@@ -2269,118 +1912,7 @@ impl Analysis {
     }
 }
 
-// Conversion functions from analysis types to IDE types
-
-/// Adjust a position for line offset (used for extracted GraphQL from TypeScript/JavaScript)
-///
-/// When GraphQL is extracted from TypeScript/JavaScript files, the line numbers in the
-/// LSP request are relative to the original file, but we need positions relative to the
-/// extracted GraphQL. This function subtracts the `line_offset` to get the correct position.
-#[allow(clippy::cast_possible_truncation)] // Line offset won't exceed u32::MAX
-const fn adjust_position_for_line_offset(position: Position, line_offset: u32) -> Option<Position> {
-    if line_offset == 0 {
-        return Some(position);
-    }
-
-    // Subtract line_offset from the incoming position
-    // If the position is before the GraphQL block, return None
-    if position.line < line_offset {
-        return None;
-    }
-
-    Some(Position::new(
-        position.line - line_offset,
-        position.character,
-    ))
-}
-
-/// Add line offset to a range (used when returning positions from extracted GraphQL)
-///
-/// When returning positions for document symbols in TypeScript/JavaScript files,
-/// we need to add the `line_offset` to convert from GraphQL-relative positions
-/// back to original file positions.
-const fn adjust_range_for_line_offset(range: Range, line_offset: u32) -> Range {
-    if line_offset == 0 {
-        return range;
-    }
-
-    Range::new(
-        Position::new(range.start.line + line_offset, range.start.character),
-        Position::new(range.end.line + line_offset, range.end.character),
-    )
-}
-
-/// Convert IDE position to byte offset using `LineIndex`
-fn position_to_offset(line_index: &graphql_syntax::LineIndex, position: Position) -> Option<usize> {
-    let line_start = line_index.line_start(position.line as usize)?;
-    Some(line_start + position.character as usize)
-}
-
-/// Convert byte offset to IDE Position using `LineIndex`
-#[allow(clippy::cast_possible_truncation)] // Line and column numbers won't exceed u32::MAX
-fn offset_to_position(line_index: &graphql_syntax::LineIndex, offset: usize) -> Position {
-    let line = line_index.line_col(offset).0;
-    let line_start = line_index.line_start(line).unwrap_or(0);
-    let character = offset - line_start;
-    Position::new(line as u32, character as u32)
-}
-
-/// Convert byte offset range to IDE Range using `LineIndex`
-fn offset_range_to_range(
-    line_index: &graphql_syntax::LineIndex,
-    start_offset: usize,
-    end_offset: usize,
-) -> Range {
-    let start = offset_to_position(line_index, start_offset);
-    let end = offset_to_position(line_index, end_offset);
-    Range::new(start, end)
-}
-
-/// Convert analysis Position to IDE Position
-const fn convert_position(pos: graphql_analysis::Position) -> Position {
-    Position {
-        line: pos.line,
-        character: pos.character,
-    }
-}
-
-/// Convert analysis `DiagnosticRange` to IDE Range
-const fn convert_range(range: graphql_analysis::DiagnosticRange) -> Range {
-    Range {
-        start: convert_position(range.start),
-        end: convert_position(range.end),
-    }
-}
-
-/// Convert analysis Severity to IDE `DiagnosticSeverity`
-const fn convert_severity(severity: graphql_analysis::Severity) -> DiagnosticSeverity {
-    match severity {
-        graphql_analysis::Severity::Error => DiagnosticSeverity::Error,
-        graphql_analysis::Severity::Warning => DiagnosticSeverity::Warning,
-        graphql_analysis::Severity::Info => DiagnosticSeverity::Information,
-    }
-}
-
-/// Convert analysis Diagnostic to IDE Diagnostic
-fn convert_diagnostic(diag: &graphql_analysis::Diagnostic) -> Diagnostic {
-    Diagnostic {
-        range: convert_range(diag.range),
-        severity: convert_severity(diag.severity),
-        message: diag.message.to_string(),
-        code: diag.code.as_ref().map(ToString::to_string),
-        source: diag.source.to_string(),
-    }
-}
-
-/// Result of finding which block contains a position
-struct BlockContext<'a> {
-    /// The syntax tree for the block (or main document)
-    tree: &'a apollo_parser::SyntaxTree,
-    /// Line offset to add when returning positions (0 for pure GraphQL files)
-    line_offset: u32,
-    /// The block source for building `LineIndex` (None for pure GraphQL files)
-    block_source: Option<&'a str>,
-}
+// Helper functions are now in helpers.rs module
 
 /// Get field children for a type definition
 fn get_field_children(
@@ -2427,597 +1959,11 @@ fn get_field_children(
     children
 }
 
-/// Find which GraphQL block contains the given position
-///
-/// For pure GraphQL files, returns the main tree with `line_offset` from metadata.
-/// For TS/JS files, finds the block containing the position and returns it with
-/// the appropriate line offset.
-#[allow(clippy::cast_possible_truncation)]
-fn find_block_for_position(
-    parse: &graphql_syntax::Parse,
-    position: Position,
-    metadata_line_offset: u32,
-) -> Option<(BlockContext<'_>, Position)> {
-    // If no blocks, this is a pure GraphQL file - use main tree
-    if parse.blocks.is_empty() {
-        let adjusted_pos = adjust_position_for_line_offset(position, metadata_line_offset)?;
-        return Some((
-            BlockContext {
-                tree: &parse.tree,
-                line_offset: metadata_line_offset,
-                block_source: None,
-            },
-            adjusted_pos,
-        ));
-    }
-
-    // For TS/JS files, find which block contains the position
-    for block in &parse.blocks {
-        let block_start_line = block.line as u32;
-        let block_start_col = block.column as u32;
-        // Calculate block end line by counting newlines in source
-        let block_lines = block.source.chars().filter(|&c| c == '\n').count() as u32;
-        let block_end_line = block_start_line + block_lines;
-
-        if position.line >= block_start_line && position.line <= block_end_line {
-            // Position is within this block
-            // Adjust position to be relative to block start (subtract both line and column)
-            let adjusted_line = position.line - block_start_line;
-            let adjusted_col = if adjusted_line == 0 {
-                // Only subtract column offset on the first line of the block
-                position.character.saturating_sub(block_start_col)
-            } else {
-                position.character
-            };
-            let adjusted_pos = Position::new(adjusted_line, adjusted_col);
-
-            return Some((
-                BlockContext {
-                    tree: &block.tree,
-                    line_offset: block_start_line,
-                    block_source: Some(&block.source),
-                },
-                adjusted_pos,
-            ));
-        }
-    }
-
-    None
-}
-
-/// Find a fragment definition in a parsed file, handling TS/JS blocks correctly
-///
-/// For pure GraphQL files, searches the main tree.
-/// For TS/JS files, searches each block and returns the correct position with block offset applied.
-#[allow(clippy::cast_possible_truncation)]
-fn find_fragment_definition_in_parse(
-    parse: &graphql_syntax::Parse,
-    fragment_name: &str,
-    content: graphql_db::FileContent,
-    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
-    metadata_line_offset: u32,
-) -> Option<Range> {
-    // For pure GraphQL files, search the main tree
-    if parse.blocks.is_empty() {
-        if let Some((start_offset, end_offset)) =
-            find_fragment_definition_range(&parse.tree, fragment_name)
-        {
-            let file_line_index = graphql_syntax::line_index(db, content);
-            let range = offset_range_to_range(&file_line_index, start_offset, end_offset);
-            return Some(adjust_range_for_line_offset(range, metadata_line_offset));
-        }
-        return None;
-    }
-
-    for block in &parse.blocks {
-        if let Some((start_offset, end_offset)) =
-            find_fragment_definition_range(&block.tree, fragment_name)
-        {
-            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
-            let range = offset_range_to_range(&block_line_index, start_offset, end_offset);
-
-            return Some(adjust_range_for_line_offset(range, block.line as u32));
-        }
-    }
-
-    None
-}
-
-/// Find a type definition in a parsed file, handling TS/JS blocks correctly
-#[allow(clippy::cast_possible_truncation)]
-fn find_type_definition_in_parse(
-    parse: &graphql_syntax::Parse,
-    type_name: &str,
-    content: graphql_db::FileContent,
-    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
-    metadata_line_offset: u32,
-) -> Option<Range> {
-    // For pure GraphQL files, search the main tree
-    if parse.blocks.is_empty() {
-        if let Some((start_offset, end_offset)) = find_type_definition_range(&parse.tree, type_name)
-        {
-            let file_line_index = graphql_syntax::line_index(db, content);
-            let range = offset_range_to_range(&file_line_index, start_offset, end_offset);
-            return Some(adjust_range_for_line_offset(range, metadata_line_offset));
-        }
-        return None;
-    }
-
-    // For TS/JS files, search each block
-    for block in &parse.blocks {
-        if let Some((start_offset, end_offset)) = find_type_definition_range(&block.tree, type_name)
-        {
-            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
-            let range = offset_range_to_range(&block_line_index, start_offset, end_offset);
-            return Some(adjust_range_for_line_offset(range, block.line as u32));
-        }
-    }
-
-    None
-}
-
-/// Find all fragment spreads in a parsed file, handling TS/JS blocks correctly
-///
-/// Returns a vector of ranges with correct positions for each block.
-#[allow(clippy::cast_possible_truncation)]
-fn find_fragment_spreads_in_parse(
-    parse: &graphql_syntax::Parse,
-    fragment_name: &str,
-    content: graphql_db::FileContent,
-    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
-    metadata_line_offset: u32,
-) -> Vec<Range> {
-    let mut results = Vec::new();
-
-    // For pure GraphQL files, search the main tree
-    if parse.blocks.is_empty() {
-        if let Some(offsets) = find_fragment_spreads(&parse.tree, fragment_name) {
-            let file_line_index = graphql_syntax::line_index(db, content);
-            for offset in offsets {
-                let end_offset = offset + fragment_name.len();
-                let range = offset_range_to_range(&file_line_index, offset, end_offset);
-                results.push(adjust_range_for_line_offset(range, metadata_line_offset));
-            }
-        }
-        return results;
-    }
-
-    // For TS/JS files, search each block
-    for block in &parse.blocks {
-        if let Some(offsets) = find_fragment_spreads(&block.tree, fragment_name) {
-            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
-            for offset in offsets {
-                let end_offset = offset + fragment_name.len();
-                let range = offset_range_to_range(&block_line_index, offset, end_offset);
-                results.push(adjust_range_for_line_offset(range, block.line as u32));
-            }
-        }
-    }
-
-    results
-}
-
-/// Find all type references in a parsed file, handling TS/JS blocks correctly
-#[allow(clippy::cast_possible_truncation)]
-fn find_type_references_in_parse(
-    parse: &graphql_syntax::Parse,
-    type_name: &str,
-    content: graphql_db::FileContent,
-    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
-    metadata_line_offset: u32,
-) -> Vec<Range> {
-    let mut results = Vec::new();
-
-    // For pure GraphQL files, search the main tree
-    if parse.blocks.is_empty() {
-        if let Some(offsets) = find_type_references_in_tree(&parse.tree, type_name) {
-            let file_line_index = graphql_syntax::line_index(db, content);
-            for offset in offsets {
-                let end_offset = offset + type_name.len();
-                let range = offset_range_to_range(&file_line_index, offset, end_offset);
-                results.push(adjust_range_for_line_offset(range, metadata_line_offset));
-            }
-        }
-        return results;
-    }
-
-    // For TS/JS files, search each block
-    for block in &parse.blocks {
-        if let Some(offsets) = find_type_references_in_tree(&block.tree, type_name) {
-            let block_line_index = graphql_syntax::LineIndex::new(&block.source);
-            for offset in offsets {
-                let end_offset = offset + type_name.len();
-                let range = offset_range_to_range(&block_line_index, offset, end_offset);
-                results.push(adjust_range_for_line_offset(range, block.line as u32));
-            }
-        }
-    }
-
-    results
-}
-
-/// Find field usages in a parsed file that match the given type and field name
-#[allow(clippy::cast_possible_truncation)]
-fn find_field_usages_in_parse(
-    parse: &graphql_syntax::Parse,
-    type_name: &str,
-    field_name: &str,
-    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
-    content: graphql_db::FileContent,
-    db: &dyn graphql_syntax::GraphQLSyntaxDatabase,
-    metadata_line_offset: u32,
-) -> Vec<Range> {
-    let mut results = Vec::new();
-
-    // For pure GraphQL files, search the main tree
-    if parse.blocks.is_empty() {
-        let file_line_index = graphql_syntax::line_index(db, content);
-        let ranges = find_field_usages_in_tree(&parse.tree, type_name, field_name, schema_types);
-        for (start, end) in ranges {
-            let range = offset_range_to_range(&file_line_index, start, end);
-            results.push(adjust_range_for_line_offset(range, metadata_line_offset));
-        }
-        return results;
-    }
-
-    // For TS/JS files, search each block
-    for block in &parse.blocks {
-        let block_line_index = graphql_syntax::LineIndex::new(&block.source);
-        let ranges = find_field_usages_in_tree(&block.tree, type_name, field_name, schema_types);
-        for (start, end) in ranges {
-            let range = offset_range_to_range(&block_line_index, start, end);
-            results.push(adjust_range_for_line_offset(range, block.line as u32));
-        }
-    }
-
-    results
-}
-
-/// Check if `current_type` matches `target_type` directly or implements it as an interface
-fn type_matches_or_implements(
-    current_type: &str,
-    target_type: &str,
-    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
-) -> bool {
-    if current_type == target_type {
-        return true;
-    }
-    if let Some(type_def) = schema_types.get(current_type) {
-        type_def
-            .implements
-            .iter()
-            .any(|i| i.as_ref() == target_type)
-    } else {
-        false
-    }
-}
-
-/// Find all field usages in a tree that match the given type and field name
-#[allow(clippy::too_many_lines)]
-fn find_field_usages_in_tree(
-    tree: &apollo_parser::SyntaxTree,
-    target_type: &str,
-    target_field: &str,
-    schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
-) -> Vec<(usize, usize)> {
-    use apollo_parser::cst::{CstNode, Definition, Selection};
-
-    fn search_selection_set(
-        selection_set: &apollo_parser::cst::SelectionSet,
-        current_type: &str,
-        target_type: &str,
-        target_field: &str,
-        schema_types: &std::collections::HashMap<std::sync::Arc<str>, graphql_hir::TypeDef>,
-        results: &mut Vec<(usize, usize)>,
-    ) {
-        for selection in selection_set.selections() {
-            match selection {
-                Selection::Field(field) => {
-                    if let Some(name) = field.name() {
-                        let field_name = name.text();
-
-                        if type_matches_or_implements(current_type, target_type, schema_types)
-                            && field_name == target_field
-                        {
-                            let range = name.syntax().text_range();
-                            results.push((range.start().into(), range.end().into()));
-                        }
-
-                        // Recurse into nested selection sets with the field's return type
-                        if let Some(nested) = field.selection_set() {
-                            if let Some(type_def) = schema_types.get(current_type) {
-                                if let Some(field_def) = type_def
-                                    .fields
-                                    .iter()
-                                    .find(|f| f.name.as_ref() == field_name)
-                                {
-                                    let field_type = field_def.type_ref.name.as_ref();
-                                    search_selection_set(
-                                        &nested,
-                                        field_type,
-                                        target_type,
-                                        target_field,
-                                        schema_types,
-                                        results,
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-                Selection::InlineFragment(inline_frag) => {
-                    let fragment_type = inline_frag
-                        .type_condition()
-                        .and_then(|tc| tc.named_type())
-                        .and_then(|nt| nt.name())
-                        .map_or_else(|| current_type.to_string(), |n| n.text().to_string());
-
-                    if let Some(nested) = inline_frag.selection_set() {
-                        search_selection_set(
-                            &nested,
-                            &fragment_type,
-                            target_type,
-                            target_field,
-                            schema_types,
-                            results,
-                        );
-                    }
-                }
-                Selection::FragmentSpread(_) => {
-                    // Fragment definitions are searched separately in the main loop,
-                    // so fields inside fragments will be found when we iterate over
-                    // FragmentDefinition. We don't need to follow the spread here.
-                }
-            }
-        }
-    }
-
-    let mut results = Vec::new();
-    let doc = tree.document();
-
-    for definition in doc.definitions() {
-        match definition {
-            Definition::OperationDefinition(op) => {
-                let root_type = match op.operation_type() {
-                    Some(op_type) if op_type.mutation_token().is_some() => "Mutation",
-                    Some(op_type) if op_type.subscription_token().is_some() => "Subscription",
-                    _ => "Query",
-                };
-
-                if let Some(selection_set) = op.selection_set() {
-                    search_selection_set(
-                        &selection_set,
-                        root_type,
-                        target_type,
-                        target_field,
-                        schema_types,
-                        &mut results,
-                    );
-                }
-            }
-            Definition::FragmentDefinition(frag) => {
-                let fragment_type = frag
-                    .type_condition()
-                    .and_then(|tc| tc.named_type())
-                    .and_then(|nt| nt.name())
-                    .map(|n| n.text().to_string());
-
-                if let (Some(fragment_type), Some(selection_set)) =
-                    (fragment_type, frag.selection_set())
-                {
-                    search_selection_set(
-                        &selection_set,
-                        &fragment_type,
-                        target_type,
-                        target_field,
-                        schema_types,
-                        &mut results,
-                    );
-                }
-            }
-            _ => {}
-        }
-    }
-
-    results
-}
-
-/// Find variable definition in an operation by name
-fn find_variable_definition_in_tree(
-    tree: &apollo_parser::SyntaxTree,
-    var_name: &str,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-) -> Option<Range> {
-    use apollo_parser::cst::{CstNode, Definition};
-
-    let doc = tree.document();
-    for definition in doc.definitions() {
-        if let Definition::OperationDefinition(op) = definition {
-            if let Some(var_defs) = op.variable_definitions() {
-                for var_def in var_defs.variable_definitions() {
-                    if let Some(variable) = var_def.variable() {
-                        if let Some(name) = variable.name() {
-                            if name.text() == var_name {
-                                let range = name.syntax().text_range();
-                                let start: usize = range.start().into();
-                                let end: usize = range.end().into();
-                                let pos_range = offset_range_to_range(line_index, start, end);
-                                return Some(adjust_range_for_line_offset(pos_range, line_offset));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Find operation definition by name
-fn find_operation_definition_in_tree(
-    tree: &apollo_parser::SyntaxTree,
-    op_name: &str,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-) -> Option<Range> {
-    use apollo_parser::cst::{CstNode, Definition};
-
-    let doc = tree.document();
-    for definition in doc.definitions() {
-        if let Definition::OperationDefinition(op) = definition {
-            if let Some(name) = op.name() {
-                if name.text() == op_name {
-                    let range = name.syntax().text_range();
-                    let start: usize = range.start().into();
-                    let end: usize = range.end().into();
-                    let pos_range = offset_range_to_range(line_index, start, end);
-                    return Some(adjust_range_for_line_offset(pos_range, line_offset));
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Find argument definition in schema type's field
-fn find_argument_definition_in_tree(
-    tree: &apollo_parser::SyntaxTree,
-    type_name: &str,
-    field_name: &str,
-    arg_name: &str,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-) -> Option<Range> {
-    use apollo_parser::cst::{CstNode, Definition};
-
-    let doc = tree.document();
-    for definition in doc.definitions() {
-        let (name_node, fields_def) = match &definition {
-            Definition::ObjectTypeDefinition(obj) => (obj.name(), obj.fields_definition()),
-            Definition::InterfaceTypeDefinition(iface) => (iface.name(), iface.fields_definition()),
-            _ => continue,
-        };
-
-        let Some(name) = name_node else { continue };
-        if name.text() != type_name {
-            continue;
-        }
-
-        let Some(fields) = fields_def else { continue };
-        for field in fields.field_definitions() {
-            let Some(fname) = field.name() else { continue };
-            if fname.text() != field_name {
-                continue;
-            }
-
-            // Found the field, now find the argument
-            if let Some(args_def) = field.arguments_definition() {
-                for input_val in args_def.input_value_definitions() {
-                    if let Some(aname) = input_val.name() {
-                        if aname.text() == arg_name {
-                            let range = aname.syntax().text_range();
-                            let start: usize = range.start().into();
-                            let end: usize = range.end().into();
-                            let pos_range = offset_range_to_range(line_index, start, end);
-                            return Some(adjust_range_for_line_offset(pos_range, line_offset));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Find the field name at a given offset (for argument context)
-fn find_field_name_at_offset(
-    tree: &apollo_parser::SyntaxTree,
-    byte_offset: usize,
-) -> Option<String> {
-    use apollo_parser::cst::{CstNode, Definition, Selection};
-
-    fn check_selection_set(
-        selection_set: &apollo_parser::cst::SelectionSet,
-        byte_offset: usize,
-    ) -> Option<String> {
-        for selection in selection_set.selections() {
-            if let Selection::Field(field) = selection {
-                let range = field.syntax().text_range();
-                let start: usize = range.start().into();
-                let end: usize = range.end().into();
-
-                if byte_offset >= start && byte_offset <= end {
-                    if let Some(args) = field.arguments() {
-                        let args_range = args.syntax().text_range();
-                        let args_start: usize = args_range.start().into();
-                        let args_end: usize = args_range.end().into();
-                        if byte_offset >= args_start && byte_offset <= args_end {
-                            return field.name().map(|n| n.text().to_string());
-                        }
-                    }
-
-                    if let Some(nested) = field.selection_set() {
-                        if let Some(name) = check_selection_set(&nested, byte_offset) {
-                            return Some(name);
-                        }
-                    }
-                }
-            }
-        }
-        None
-    }
-
-    let doc = tree.document();
-    for definition in doc.definitions() {
-        match definition {
-            Definition::OperationDefinition(op) => {
-                if let Some(selection_set) = op.selection_set() {
-                    if let Some(name) = check_selection_set(&selection_set, byte_offset) {
-                        return Some(name);
-                    }
-                }
-            }
-            Definition::FragmentDefinition(frag) => {
-                if let Some(selection_set) = frag.selection_set() {
-                    if let Some(name) = check_selection_set(&selection_set, byte_offset) {
-                        return Some(name);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    None
-}
-
-/// Format a type reference for display (e.g., "[String!]!")
-/// Unwrap a `TypeRef` to get just the base type name (without List or `NonNull` wrappers)
-fn unwrap_type_to_name(type_ref: &graphql_hir::TypeRef) -> String {
-    type_ref.name.to_string()
-}
-
-fn format_type_ref(type_ref: &graphql_hir::TypeRef) -> String {
-    let mut result = type_ref.name.to_string();
-
-    if type_ref.is_list {
-        result = format!("[{result}]");
-        if type_ref.inner_non_null {
-            result = format!("[{}!]", type_ref.name);
-        }
-    }
-
-    if type_ref.is_non_null {
-        result.push('!');
-    }
-
-    result
-}
-
 #[cfg(test)]
 #[allow(clippy::needless_raw_string_hashes)]
 mod tests {
     use super::*;
+    use crate::helpers::{convert_position, convert_range, convert_severity};
 
     #[test]
     fn test_analysis_host_creation() {
